backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0
sqladmin==0.21.0
wtforms==3.1.2
alembic==1.12.1
itsdangerous==2.1.2
sendgrid
pdfkit
wkhtmltopdf
requests
